Backend/requirements.txt
asgiref==3.10.0
Django==5.2.8
sqlparse==0.5.3
tzdata==2025.2
djangorestframework==3.15.2
requests==2.32.3
